APplied fix on URL's containing an & and where it would be mishandled by the background audio player. In order to fix that, we are using an URL shortening service so we don't have &'s in the AudioTrack's  source URI.
In order to do that with Silverlight's threads handling in http requests, It was needed to change the AudioTrack factory to a proper async Create method call. If no & is found in the URI, then the asynchronous callback is called directly and on the same thread before even returning from the call to the Create method of the factory.